+++ integration-tests/emulated/tests/collectives/collectives-polkadot/src/tests/mod.rs
@@ -13,4 +13,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod fellowship_salary;
 mod fellowship_treasury;